Updates and better install.rb

+++ install.rb
@@ -2,14 +2,24 @@
 
 home = File.expand_path('~')
 
-Dir['*'].each do |file|
-  next if file =~ /install/
-  target = File.join(home, ".#{file}")
-  `ln -s #{File.expand_path file} #{target}`
+
+def symlink_files
+  Dir['*'].each do |file|
+    next if file =~ /install/
+    target = File.join(home, ".#{file}")
+    unless File.exists? target
+      `ln -s #{File.expand_path file} #{target}`
+    end
+  end
+end
+
+def make_vim_tmp
+  vim_tmp = File.join(home, '.vim-tmp')
+  unless File.exists? vim_tmp
+    `mkdir #{vim_tmp}`
+  end
 end
 
 
-vim_tmp = File.join(home, '.vim-tmp')
-unless File.exists? vim_tmp
-  `mkdir #{vim_tmp}`
-end
+symlink_files
+make_vim_tmp
